requirements.txt
# requirements.txt (VERSÃO ESTÁVEL)
Flask
google-api-python-client
google-auth
b2sdk